requirements.txt
# requirements.txt
fastapi
uvicorn
pydantic
langchain-google-genai
langgraph
python-dotenv
langchain-groq